Added Description and Main Function
Completed research on the issue that was trying to be solved with this code and did research to supplement the parts of my code that needed it. I also copy and pasted my main function from my last project and edited the variable names, the input messages, and some of the user experience formatting in it. I set up all the functions I need and put print statements in them so I can ensure the main function calls them when needed. I also created a sample file to use in one of the functions.