Trial and Error Process
Tried some code related to user experience. Also retrieved the contents in a file and stored it in a dictionary. The dictionary was organized based on what type of food it is and then checked if it has the proper amount of nutrition. The matplotlib package was also used to create a chart with the nutrition value. The budget and total cost was also added up and stored to print out.