User Experience Progress
Added a table function to format the text in a neater, understandable way. Also worked on error handling for integers, floats, and other strings through the use of try and except. I worked the most on changing the print statement by adding symbols and emojis to make the console more appealing and less boring. Fixed an error happening with the cost variable, where it was staying as a string because I was using string format to make it have only 2 decimals. But I fixed this by first converting the variable to a float and doing any of the math and other processes that require it to be a float, and then I changed it to a format string with two decimal places for visual appearance and storing in lists.