Programmed the Add History and Show history Functions
Stored the user's score in each type of food, along with their budget and total cost, and their numerical scores in a file (which is in append mode). Then created the show history function which reads the file and prints out the previous dietary info stated above by its date. Also worked a little on user experience in all functions and is running into some errors regarding calling functions and error handling function.